Added Discretization Scheme to the interface
View now references MonteCarlo.Model and databinds a new combobox -
cbDiscretizationScheme to a list of possible schemes when it starts.
Presenter
now sets the current scheme in the RunSimulation method, so the
constructor of
the Simulator class has been removed.

Minor formatting changes have been made to the Simulator.cs file to make
it
easier to read.